fix bloat16 conversion issue

+++ gbx_lm/utils.py
@@ -834,8 +834,10 @@ def load_model(
         for k, v in weights.items():
             if 'qweight' in k:
                 weights[k] = v.transpose().astype(mx.uint32)
+                
             if not use_double_quantization and ('scales' in k or 'zeros' in k):
-                weights[k] = v.transpose()
+                weights[k] = v.transpose().astype(mx.bfloat16)
+
             if "norm.weight" in k or "bias" in k or "gate.weight" in k or "lm_head" in k or "embed_tokens" in k or "channel_scale" in k:
                 weights[k] = v.astype(mx.bfloat16)
     ## ===================================================================##
